// src/common/pagination/providers/pagination.provider.ts
import { Inject, Injectable } from '@nestjs/common';
import { ObjectLiteral, Repository } from 'typeorm';
import { paginationQueryDto } from '../dtos/pagination-query.dto';
import { REQUEST } from '@nestjs/core';
import type { Request } from 'express';
import { Paginated } from '../interfaces/paginated.interface';

@Injectable()
export class PaginationProvider {
    constructor(
        @Inject(REQUEST)
        private readonly request: Request, 
    ){}

    public async paginateQuery<T extends ObjectLiteral>(
        paginateQuery: paginationQueryDto, 
        Repository: Repository<T>
    ): Promise<Paginated<T>>{
        let results = await Repository.find({ 
            skip: (paginateQuery.page - 1) * paginateQuery.limit,
            take: paginateQuery.limit,
        });

        const baseURL = this.request.protocol + '://' + this.request.headers.host + '/';
        const newUrl = new URL(this.request.url, baseURL);
        
        //calculating page number
        const totalItems = await Repository.count();
        const totalPages = Math.ceil(totalItems/paginateQuery.limit);
        const nextPage = paginateQuery.page === totalPages 
            ? paginateQuery.page 
            : paginateQuery.page + 1;

        const previousPage = paginateQuery.page === 1 
            ? paginateQuery.page 
            : paginateQuery.page - 1;

        const finalResponse: Paginated<T> = {
            data: results,
            meta:{
                itemsPerPage:paginateQuery.limit,
                totalItems: totalItems,
                currentPage: paginateQuery.page,
                totalPages: totalPages,
            },
            links:{
                first: `${newUrl.origin}${newUrl.pathname}?limit=${paginateQuery.limit}&page=1`,
                last: `${newUrl.origin}${newUrl.pathname}?limit=${paginateQuery.limit}&page=${totalPages}`,
                current: `${newUrl.origin}${newUrl.pathname}?limit=${paginateQuery.limit}&page=${paginateQuery.page}`,
                next: `${newUrl.origin}${newUrl.pathname}?limit=${paginateQuery.limit}&page=${nextPage}`,
                previous: `${newUrl.origin}${newUrl.pathname}?limit=${paginateQuery.limit}&page=${previousPage}`,
            },

        };
        // console.log(baseURL);
        // console.log(newUrl);

        return finalResponse;
    }
}
